修复休息状态时点击提前完成导致重复记录的BUG

+++ app/tool4time.py
@@ -57,11 +57,11 @@ def get_timestamp_from_str(t: str) -> float:
 
 # ################################# API For Server ################################# #
 
-def parse_timestamp(timestamp: int) -> str:
+def parse_timestamp(timestamp: float) -> str:
     return time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(timestamp))
 
 
-def parse_deadline_timestamp(timestamp: int) -> str:
+def parse_deadline_timestamp(timestamp: float) -> str:
     return time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(timestamp / 1000))
